Adjusting Test project to separate namespace for Integration tests and unit tests
