+++ Marajoara.Cinema.Management/Marajoara.Cinema.Management.Tests/Integration/Repositories/MovieRepositoryTests.cs
@@ -1,11 +1,11 @@
 ﻿using FluentAssertions;
 using Marajoara.Cinema.Management.Domain.MovieModule;
-using Marajoara.Cinema.Management.Tests.Common;
+using Marajoara.Cinema.Management.Tests.Integration.Common;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Marajoara.Cinema.Management.Tests
+namespace Marajoara.Cinema.Management.Tests.Integration.Repositories
 {
     [TestClass]
     public class MovieRepositoryTests : UnitOfWorkIntegrationBase
